+++ gltf/src/net/mgsx/gltf/shaders/skybox.vs.glsl
@@ -1,10 +1,8 @@
-attribute vec3 a_position;
-uniform mat4 u_projViewTrans;
-uniform mat4 u_worldTrans;
-varying vec3 v_position;
+attribute vec4 a_position;
+varying vec4 v_position;
 
 void main() {
-	v_position = a_position.xyz;
-	vec4 pos = u_worldTrans * vec4(a_position, 1.0);
-	gl_Position = u_projViewTrans * pos;
+	v_position = a_position;
+	gl_Position = a_position;
+	gl_Position.z = 1.0;
 }